Clean up and move new test files
